requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
psycopg2-binary==2.9.9
python-slugify==8.0.4
pyotp==2.9.0
gunicorn==22.0.0
